implemented new approach to get the campaigns that the user donated to
