+++ apps/api/src/campaign/campaign.service.ts
@@ -33,7 +33,7 @@ export class CampaignService {
     private prisma: PrismaService,
     @Inject(forwardRef(() => VaultService)) private vaultService: VaultService,
     @Inject(forwardRef(() => PersonService)) private personService: PersonService,
-  ) {}
+  ) { }
 
   async listCampaigns(): Promise<Campaign[]> {
     const campaigns = await this.prisma.campaign.findMany({
@@ -149,6 +149,62 @@ export class CampaignService {
     const campaignSums = await this.getCampaignSums([campaign.id])
 
     return this.addVaultAndDonationSummaries(campaign, campaignSums)
+  }
+
+  async getUserCampaigns(id: string): Promise<Campaign[]> {
+    const campaigns = await this.prisma.campaign.findMany({
+      where: {
+        OR: [{ beneficiary: { person: { id: id } } }, { coordinator: { person: { id: id } } }, { organizer: { person: { id: id } } }]
+      },
+      orderBy: {
+        endDate: 'asc',
+      },
+      include: {
+        campaignType: { select: { name: true, slug: true } },
+        beneficiary: {
+          select: {
+            id: true,
+            type: true,
+            person: { select: { id: true, firstName: true, lastName: true } },
+            company: { select: { id: true, companyName: true } },
+          },
+        },
+        coordinator: { select: { person: { select: { firstName: true, lastName: true } } } },
+        organizer: { select: { person: { select: { firstName: true, lastName: true } } } },
+        incomingTransfers: { select: { amount: true } },
+        outgoingTransfers: { select: { amount: true } },
+      },
+    })
+    const campaignSums = await this.getCampaignSums()
+
+    return campaigns.map((c) => this.addVaultAndDonationSummaries(c, campaignSums))
+  }
+
+  async getUserDonatedCampaigns(keycloakId: string) {
+    const campaigns = await this.prisma.campaign.findMany({
+      where: { vaults: { some: { donations: { some: { person: { keycloakId } } } } } },
+      orderBy: {
+        endDate: 'asc',
+      },
+      include: {
+        campaignType: { select: { name: true, slug: true } },
+        beneficiary: {
+          select: {
+            id: true,
+            type: true,
+            person: { select: { id: true, firstName: true, lastName: true } },
+            company: { select: { id: true, companyName: true } },
+          },
+        },
+        coordinator: { select: { person: { select: { firstName: true, lastName: true } } } },
+        organizer: { select: { person: { select: { firstName: true, lastName: true } } } },
+        incomingTransfers: { select: { amount: true } },
+        outgoingTransfers: { select: { amount: true } },
+      },
+    })
+    const campaignSums = await this.getCampaignSums()
+
+    return campaigns.map((c) => this.addVaultAndDonationSummaries(c, campaignSums))
   }
 
   async getCampaignByIdAndCoordinatorId(
@@ -341,9 +397,9 @@ export class CampaignService {
     const vault = await this.prisma.vault.findFirst({ where: { campaignId } })
     const targetVaultData = vault
       ? // Connect the existing vault to this donation
-        { connect: { id: vault.id } }
+      { connect: { id: vault.id } }
       : // Create new vault for the campaign
-        { create: { campaignId, currency: campaign.currency, name: campaign.title } }
+      { create: { campaignId, currency: campaign.currency, name: campaign.title } }
 
     // Find donation by extPaymentIntentId and update if status allows
 
@@ -356,9 +412,9 @@ export class CampaignService {
     if (!donation) {
       Logger.debug(
         'No donation exists with extPaymentIntentId: ' +
-          paymentData.paymentIntentId +
-          ' Creating new donation with status: ' +
-          newDonationStatus,
+        paymentData.paymentIntentId +
+        ' Creating new donation with status: ' +
+        newDonationStatus,
       )
 
       try {
